Modified the output where link to the effect's pointer is preserved
Now output includes as the first column the original reference
number for the effect in move_effect_prose.csv. Basically the
original csv file has been split into multiple lines based
on if the short effect description describes a combined effect.

This helps both exploring the domain and also in future parsing
the data itself because the link to the original data set is
still preserved in this model.