Move effect prose parsing rewritten; all effects on single lines
Each atomic effect is now on a single line, and the conncection
to the effect_id has been preserved, so the new csv file can
be used to both model the overall language of Pokemon effects,
and also make it easier to parse the moves themselves further
in the process.

The whole thing is done in a shell script parse_move_effect_prose.sh,
and it uses the new move_effect-prefixed AWK-scripts in certain
situations. The old separate_individual_effects script is thus
redundant and will be deleted in the future.