AWK script for categorizing move effects put into effect
This script is still incomplete and it will be filled later on.
Currently all effects lasting n turns and effects which happen
only after if-condition are handled. Needs further refinement
on separating effects, which can affect any Pokemon on the field
from those that are purely user or target specific.